requirements.txt
requests
webull
pyyaml
pandas
schedule
pytest
pytest-cov
